pyproject.toml
[project]
name = "qs-migration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.36.11",
    "icecream>=2.1.4",
    "loguru>=0.7.3",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "boto3-stubs[full]>=1.36.11",
]
